Move light normalization logic to new processor
